HW/requirements.txt
apache-airflow-providers-snowflake==5.6.0
snowflake-connector-python==3.11.0
